Make private _globals and _constant module as torch 1.12
